animwal/pyproject.toml
[tool.poetry]
name = "animwal"
version = "0.1.1"
readme = "README.md"
description = "animated wallpapers"
authors = ["nwxnk <nwxnk@yandex.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pysdl2 = "^0.9.7"
pillow = "^7.2.0"

[tool.poetry.scripts]
animwal = "animwal.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"